+++ CHANGELOG.md
@@ -4,10 +4,16 @@
 
 ### Fixed
 ### Changes
+- Refactoring: Moved all extension functions and properties targeting `TreeView`
+  from `Nodes.kt` to `TreeView.kt`. 
+  
+  **Note:** This could break imports in your kt files, please reimport or use `no.tornadofx.*`!
+  
 ### Additions
 
 - menubutton builder (https://github.com/edvin/tornadofx/issues/461)
 - MenuButton.item builder
+- Added Fragment support for`TreeCell`
 
 ## [1.7.11]